Add support for the fixed_cell option to the Recharge package (#1005)
Add support for the ``fixed_cell`` option to the Recharge package. This
option is relevant for phreatic models, not using
the Newton formulation and model cells can become inactive. The
preferred method
for phreatic models is to use the Newton formulation, where cells remain
active,
and this option irrelevant.

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [ ] Links to correct issue
- [x] Update changelog, if changes affect users
- [ ] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [x] Unit tests were added
- [ ] **If feature added**: Added/extended example