Issue #1013 Make linting stricter (#1014)
Fixes #1013

# Description
This commit makes the ruff formatter and linter stricter.
When switching to ruff we had to ignore a lot of the linting rules to
keep the build green.
Most of the ignored rules have been resolved in this commit

The rules have been grouped per commit. To help you in reviewing this
you can go through the commits one by one to see related changes

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [ ] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example