Issue #1026 Let MyPy lint more files (#1027)
Fixes #1026 

In the PyProject.toml file there is a list of glob statements which
tells MyPy which files is should lint.
There was a mistake however that caused MyPy to only lint the top level
files of a folder.
This has been fixes in this commit.

Due to more files being analysed by MyPy there were some new issues
found
Those issues are fixed in this commit as well

I tried to make the reviewing process easier by grouping related changes
to the same commit.

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [ ] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example