Issue #1034 Add Stricter MyPy settings (#1035)
Fixes #1034

# Description
This commit adds the MyPy stricter setting to the PyProject.toml file
and also fixes some more MyPy errors.

The result of this settings is that instead of rules being disabled by
default we know have to disable them explicitly.
This make it visible which rules we are not applying and we still need
to fix.

The ideal situation would be to have no mypy rules at all in tjhe
project files which means we are using complying to all MyPy rules.
The road to MyPy strict settings is described by
https://mypy.readthedocs.io/en/stable/existing_code.html#introduce-stricter-options

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [ ] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example