Issue #1041 import simulation (#1059)
Fixes #1041

# Description
Adds a function that imports a simulation from an imod5 project file. It
creates a simulation, containing 1 GroundwaterFlowModel, containing the
packages for which we support importing currently. This includes the
mandatory flow packages like NPF and STO, and some boundary conditions
(like DRN and RCH) but not yet CHD or WEL.

# Checklist
- [X] Links to correct issue
- [ ] Update changelog, if changes affect users
- [X] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [X] Unit tests were added
- [ ] **If feature added**: Added/extended example

---------

Co-authored-by: Joeri van Engelen <joerivanengelen@hotmail.com>