Issue #966 chd import or shd (#1088)
Fixes #966

# Description
Imports chd packages present in imod5_data. These packages are imported
one by one. If none are present, then a chd package is created at all
locations where ibound == -1. In that case the starting head of the
simulation is assigned as the chd value for each cell with ibound == -1.

# Checklist
- [X] Links to correct issue
- [ ] Update changelog, if changes affect users
- [X] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [X] Unit tests were added
- [ ] **If feature added**: Added/extended example