Issue #678 hfb varying top bot (#1100)
Fixes #678 

# Description
Support horizontal flow barriers with varying ztop and zbottom. Data is
set up as vertical polygons, which is also what
``imod.formats.gen.read_gen()`` returns. The implementation is based on
``imod.formats.prj.disv_converter``. When we fix the wells, we are
feature complete and the ``disv_converter`` can be removed.

The use of the vertical dimension means a change in the public API: No
longer are the "ztop" and "zbottom" variables in the geodataframe used.
Setting up these vertical polygons from scratch is a bit of a hassle, so
we could add some utility functions to help setting these up, in a later
PR.

# Checklist
- [x] Links to correct issue
- [x] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [x] Unit tests were added
- [x] **If feature added**: Added/extended example