Issue #1139 apply factor and addition (#1140)
Fixes #1139 

# Description
Loading project files turned out to be a time-sink for project file
imports- especially when loading prj-file data with transient arrays
such as chd-levels per timestep.
These arrays were loaded in a function called
"apply_factor_and_addition" which accounted for the fact that the user
can specify a factor to multiply each array and a constant to add to it.
These 2 (factor and addition) can be different for every timestep.
Now they are applied at an earlier stage, when loading the ipf data. 

No new tests were added, because the tests in test_import_prj.py already
test the application of factors and additions. (this PR is for improving
performance and not changing behavior)

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [ ] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example