Issue #1132 performance hfb (#1157)
Fixes #1132 

# Description
This PR adds the following:

- Increase performance for writing the HFB files, by using new xugrid
create_snap_to_grid_dataframe.
- Fixes error when multiple barriers in a single dataset would be
snapped to same edge. These lines are now aggregated.
- Add some basic infrastructure for user acceptance test for writing
LHM, the path to the LHM projectfile should be set in a .env as LHM_PRJ.
This can later be extended to TeamCity.

This reduces the time spent writing the HFB packages from 34 minutes to
12.5 minutes on my local machine. I've profiled the causes of the
remaining bottlenecks to writing. I'll post a follow-up issue for that.

# Checklist

- [x] Links to correct issue
- [x] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [x] Unit tests were added
- [ ] **If feature added**: Added/extended example