Issue #1187 cleanup wells (#1206)
Fixes #1187

# Description
Adds cleanup utility for wells, as well as a method on the ``Well``
object. ``LayeredWell`` doesn't have ``cleanup`` method yet, as cleaning
up almost purely focuses on adjusting filter screen levels for now. The
function as well as the method are a bit more involved than the other
cleanup utilities, because we are dealing here with a comparison of
point data and grids.

# Checklist
- [x] Links to correct issue
- [x] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [x] Unit tests were added
- [ ] **If feature added**: Added/extended example