Issue #1188 Update the numba_celltree package (#1189)
Fixes #1188

# Description
In the numba_celltree package a change has been made that changes the
behavior of the locate_points method.

In structured grids the behavior is that when a point lies on the left
or bottom edge it is considered to be in the cell.
When the point lies on the right or top edge its considered outside of
the cell

This kind of logic doesn't work for unstructured grids because the
orientation of the cell can be anything.
However there are unit tests that do assume that this applies for
unstructured grids as well. And those are the tests that are failing
right now.

To solve this the bahavior has changed such that when a point lies on a
cell edge its is considered to be in the cell. This also applies to
structred grids.


This was

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [ ] Links to correct issue
- [ ] Update changelog, if changes affect users
- [ ] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example