Issue #1150 Use default Pydata colors (#1215)
Fixes #1150

# Description
We had a lot of issues using our custom colorscheme with the pydata
theme.
Most of the times sphinx updated our colorscheme broke and we had to
manually fix it.

To avoid any headache in the future we decided to use their default
colors.
In this PR our custom colors are removed and sphinx has been updated to
the latest version

Further more changes have been made to the imod-logo and deltares logo
making it compatible with the light and dark mode

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [ ] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example

---------

Co-authored-by: Joeri van Engelen <joerivanengelen@hotmail.com>