Issue #1209 bypass value error well assign (#1211)
Fixes #1209

# Description
- Add setting to ``to_mf6_pkg``: ``error_on_well_removal``, which throws
an error if wells are removed entirely during well assignment.
- Separated code to collect filtered well ids in separate method. This
method is now also called to collect ids of errors being removed in
error message.
- Added attribute ``_is_from_imod5`` to Modflow6Simulation to keep track
if the simulation is loaded from iMOD5 data or not. Based on this, some
validation settings might be set to be a bit more permissive. (Or call
some extra cleanup?)
- ``is_from_imod5`` property to WriteContext to keep track of the same
thing within the writing context.

# Checklist
- [x] Links to correct issue
- [x] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [x] Unit tests were added
- [ ] **If feature added**: Added/extended example