Use apply_ufunc in `imod.prepare.fill`. Fixes #1237 (#1238)
Fixes #1237.

This is unfortunately a breaking change, but keeping the previous API
makes the logic a lot more complicated (and arguably makes the function
worse).

For example, keeping the `invalid` (might need a better name too) as a
DataArray almost surely requires including it in the broadcasting logic
of `apply_ufunc`.

---------

Co-authored-by: Joeri van Engelen <joerivanengelen@hotmail.com>