Issue #1263 Fix failing pixi pipeline (#1264)
Fixes #1263

# Description
Something has changed on TeamCity. The pixi pipeline is now unable to
determine the users emailadres.
This resulted the pipeline script being unable to create/push a commit.

This PR sets the username and email explicitly which should fix the
issue

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [ ] Links to correct issue
- [ ] Update changelog, if changes affect users
- [ ] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example