Issue #1255 msw refactor mf6 obj at write (#1259)
Fixes #1255 and #1056

# Description
This is part 2 of the metaswap refactoring on the iMOD Python side. The
goal of this step was to require a Mf6Wel object at ``write`` instead of
at init. For this I was struggling quite a lot with Linkov's principle
and the resulting mypy errors. In the end, I resorted to requiring all
arguments for the write method for each package. This results in a bit
of a weird situation, where a ``None`` has to be provided for the
``mf6_dis`` and ``mf6_wel`` arguments to the ``write`` of an individual
no even using these args. I don't think this is a major hickup for
users, as they are nearly always writing with ``model.write`` directly,
and not writing individual packages manually.

Though not perfect, I think this is certainly a big improvement to the
current situation.

In total, this PR:

- Removes the mf6 packages as attributes from ``Sprinkling`` and
``CouplerMapping``. This makes dumping the metaswap packages easier.
This affects the files ``imod.msw.sprinkling.py`` and
``imod.msw.coupler_mapping.py``.
- Updates the ``Sprinkling`` and ``CouplerMapping`` class to ask
MODFLOW6 packages at ``write`` instead of at ``__init__``. This affects
``imod.msw.sprinkling.py``, ``imod.msw.coupler_mapping.py``, and
``imod.msw.model.py``
- Updates the signatures
- Implements a ``regrid_like`` for the Sprinkling package. Affects
``imod.msw.sprinkling.py``.
- Adds a test to regrid a MetaSWAP model and its coupled MODFLOW6 model.
The ``Mf6Wel`` package has to be created manually now after regridding.

# Checklist
- [x] Links to correct issue
- [x] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [x] Unit tests were added
- [ ] **If feature added**: Added/extended example