Issue #1261 Support imod5 ipf data in cap for open_projectfile_data (#1265)
Fixes #1261

# Description
This PR changes the following things:
- expands ``imod.prj.open_projectfile_data`` so that it doesn't break
upon reading a CAP (with data for MetaSWAP) where an IPF is specified
for sprinkling wells, instead of a IDF.
- slightly refactors the projectfile tests, to work with two test cases,
and to move the projectfile template outside the setup. This allows
expanding the prj template and adding more test cases for the
projectfile.
- some boyscouting: move some of the logic in ``_read_package_ipf`` to
separate functions, to reducte the function's length. This is not really
that relevant to this PR, but I started working on this thinking I
needed the _read_package_ipf. It turned out this wasn't necessary.

# Checklist

- [x] Links to correct issue
- [x] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [x] Unit tests were added
- [ ] **If feature added**: Added/extended example