Fix Issue #1262 Purge np.bool, convert to np.bool_ (#1272)
Fixes #1262

# Description
``np.bool`` was an alias for the builtin bool. It was removed in numpy
>1.23, but then reintroduced in numpy 2.0 for some reason.

It was used in a type hint ``points_in_bounds``. If you try to install
iMOD Python with 1.26, it breaks upon import.

I think adding extra test jobs to for all kinds of versions for
dependencies, defeats the purpose of the test bench, so I didn't add any
extra tests. This issue is very specific, as iMOD Python now breaks for
very specific versions of numpy: >=1.23, <2.0.

# Checklist

- [x] Links to correct issue
- [x] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example