Issue #1273 cleanup example (#1274)
Fixes #1273

# Description
Add example of cleanup utilities to the user guide. I think it shows a
common use-case for cleaning up data.

Technically the most challenging issue was to work around the fact that
sphinx_gallery doesn't have an option to select expected errors in
specific code blocks. There is [this
PR](https://github.com/sphinx-gallery/sphinx-gallery/pull/930) to
introduce this, but unfortunately for us this has not been finished yet.

I resorted to trapping the error and printing it with a context manager,
that seemed like the cleanest solution to me. I added this small context
to the public API, so that users can also read in the online
documentation what the context manager does.

# Checklist

- [x] Links to correct issue
- [x] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [x] Unit tests were added
- [x] **If feature added**: Added/extended example