Issue #1183 Update documentation on package managers (#1275)
Fixes #1183 and #1185

# Description
* I did not want to remove references to Anaconda entirely: Anaconda is
still well known, so we need our docs to briefly explain what it is and
why you could use an alternative. I reduced the mentions of it.
* I added pixi as the recommended way of installing imod, since it seems
to be stable enough now. The recently added functionality for a global
installation also makes pixi behave more like users are used to from
conda.
* I removed mentions of ``mamba``, as its faster dependency solver is
now the default in ``conda``, and the lead developer moved on to develop
``pixi``. Mamba is a success in my opinion, but it has served its
purpose: Its killer feature has been moved to conda, and mamba will be
less maintained. At least I think, I haven't seen any mentions of
sunsetting it though and it has an active userbase. Regardless, three
options for package installers (pixi, conda, pip) is already
overwhelming for most of our users, so removing this
formerly-critical-now-less-useful option is necessary.

# Checklist

- [x] Links to correct issue
- [ ] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example