Issue #1192 Add the dockerfile for the docker build image (#1284)
Fixes #1192 

# Description
This PR add the dockerfile used to create the docker image which is
being used in the build pipeline.
Documentation has been added how to build and update it. Furthermore the
section about debugging a failing CI build has been extended to explain
how to run the container locally to debug it.

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [ ] Links to correct issue
- [ ] Update changelog, if changes affect users
- [ ] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example