Issue #1285 grid data from imod5 (#1286)
Fixes #1285 

# Description

- This PR implements ``GridData.from_imod5_data``, which has the
following requirements:
  - Area: Split area in rural & urban area
  - Active: Split where active/inactive
  - Landuse: Set urban_landuse to 18
  - Rootzone thickness: convert from centimeter to meter
- Refactors the ``test_grid_data.py`` to separate cases, reduces code
duplication
- Add unittest for ``GridData.from_imod5_data``

It has a arguments for regridding, but these don't do anything yet. I
wasn't sure if I should keep this to already have a future-proof
signature, or not add these, as the regridder arguments only confuse
people. I don't think regridding the MetaSWAP iMOD5 grids (like we do
with MODFLOW6 data) is a part of iMOD Python now, as we are first
focusing on the LHM, for which this is not a hard requirement.

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [ ] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [x] Unit tests were added
- [ ] **If feature added**: Added/extended example