Issue #1248 Write svat_groundwater to file and update the tests (#1249)
Fixes #1248 

# Description

This PR fixes Issue #1248 by removing the ``svat_groundwater`` from the
_to_fill variable. For now the source svat == target svat.

In addition the ``max_abstraction_groundwater`` should now contains the
subunit coordinate. The current implementation assigned the abstraction
to all stacked subunits. This is not desirable in case the second svat
represents the urban area for example. Now its up to the user to define
irrigation at the subunit level.

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [x] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [x] Unit tests were added
- [ ] **If feature added**: Added/extended example

---------

Co-authored-by: JoerivanEngelen <joerivanengelen@hotmail.com>