Issue #1313 recharge from imod5 cap data (#1315)
Fixes #1313

# Description
Construct an rch-package from iMOD5 data in the CAP package, loaded with
the ``open_projectfile_data`` function. Package is
used to couple MODFLOW6 to MetaSWAP models. Active cells will have a
recharge rate of 0.0.

At the moment, MetaSWAP can only be coupled to the first layer, as this
is also the case for ``primod``. iMOD Coupler these days supports
coupling to other layers as well, but ``primod`` doesn't. Picking this
up for iMOD Python and ``primod`` is worthy a separate story.

In detail:
- Add ``Recharge.from_imod5_cap_data`` class method, to construct an
empty Recharge package with 0.0 rate.
- Minor refactor: Put ``GridData`` helper functions to
``msw/utilities/imod5_converter.py``, so that they can be reused.

# Checklist

- [x] Links to correct issue
- [x] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [x] Unit tests were added
- [ ] **If feature added**: Added/extended example