Issue #1311 sprinkling from imod5 data (#1312)
Fixes #1311

# Description

Adds the following:

- ``imod.mf6.LayeredWell.from_imod5_cap_data``, to construct
LayeredWells for MODLOW6 from the CAP package in iMOD5 Data. Throws an
error if data is provided as point data (IPF, which is read as pandas
DataFrame). This has rate 0.0, actual rates will be inserted by in the
coupling scheme of iMOD Coupler.
- ``imod.mf6.GroundwaterFlowModel.from_imod5_data`` adds Sprinkling well
if CAP package present. This does not affect MODFLOW6 model if not
coupled to MetaSWAP: The wells have rates 0.0, so will not affect the
simulation results.
- ``Sprinkling.from_imod5_data`` to construct a Sprinkling package from
iMOD5 data.

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [x] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [x] Unit tests were added
- [ ] **If feature added**: Added/extended example