Issue #1330 imod5 docs (#1333)
Fixes #1330

# Description
Adds tables with documentation of the status on important iMOD5 features
which we support, where we support them. Things we do not support (yet?)
get a dot now.


![image](https://github.com/user-attachments/assets/a6f281f0-8cf6-4776-82bc-8ce934c6b5a0)

I put it under the "FAQ" as it is something we get questions on and to
be fair, the FAQ is the only place to conveniently store this more
detailed, not-code-example, information to users. An alternative might
be the API docs, but I wasn't entirely sure.