Issue #1383 Add support for ``ALLOCATION_OPTION.stage_to_riv_bot_drn_above`` in ``River.from_imod5_data``. (#1384)
Fixes #1383 and #1060

It turned out the differences in #1383 were mainly caused by a lack of
support of ``ALLOCATION_OPTION.stage_to_riv_bot_drn_above``, which is
the default setting for iMOD5, but we didn't support it yet.
This was because we lacked some logic to do something with the extra
addition Drainage package created by with this option during allocation.
This PR implements the following:

- Support for ``ALLOCATION_OPTION.stage_to_riv_bot_drn_above`` in
``River.from_imod5_data``.
- Set ``ALLOCATION_OPTION.stage_to_riv_bot_drn_above`` as default
allocation option for rivers
- Adapt some unittests to run with the new default option
- Add two unittests cases to ensure rivers are properly allocated and
cleaned when ``stage`` equals ``bottom_elevation``. I first thought
there was a bug in one of these functions, so I added this edge case.
But they ran without issue! Still good to keep these tests for this edge
case to ensure no mistakes creep into the code base over time.