Add python 3.13 support (#1404)
Fixes #1317 and #1331

# Description
With `Numba` now also supporting python 3.13 we can also add 3.13
support to iMOD-python.
This PR adds a 3.13 environment to the pixi.toml and also adds an import
test of it to the TC pipelines.

One line of code had to be rwritten to be 3.13 compatible. The
`importlib.resources.files` no longer provides a context manager and has
to wrapped using the `importlib.resources.as_file` method

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [ ] Links to correct issue
- [ ] Update changelog, if changes affect users
- [ ] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example