Issue #1436 work around dask issue (#1450)
Fixes #1436 

# Description
Work around Dask issue by forcing a load of idomain and svat into
memory. These grids will never be as large as the boundary condition
grids with a time axis, so I think it is acceptable. At least for the
LHM it is. I couldn't get it to work with a ``where``: numpy only
accepts 1D arrays on a boolean indexed 3D as long as the 1D array has
the length of one of the dimensions, it then does what we want to do
here. However when this is not the case, which is 95% of the time, numpy
throws an error. I therefore had to resort to loading arrays into
memory.

Furthermore I refactored ``test_coupler_mapping.py`` to be able to
reduce some code duplication and add a test case with a dask array. For
this I had to make the well package used consistent (From 2 wells to 3
wells, where the second well now lays on y=2.0 instead of y=1.0), so I
updated some of the assertions.

# Checklist
- [x] Links to correct issue
- [x] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [x] Unit tests were added
- [ ] **If feature added**: Added/extended example