Issue #1414 add cleanup hfb (#1426)
Fixes #1414 

# Description
- Adds ``cleanup_hfb`` function to clip barrier lines/Z polygons outside
the active domain. This unfortunately cannot deal with the edge case
where a hfb line is snapped to the outer edge of the active domain.
(These edges are filtered out when calling ``to_mf6_pkg``, but a
ValidationError is thrown when ``strict_hfb_validation`` is switched
on).
- Adds ``HorizontalFlowBarrierBase.cleanup`` method
- Moves some utility functions from ``imod.mf6.utilities`` to
``imod.prepare.hfb`` and ``imod.util.hfb``, to avoid circular imports.

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [x] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [x] Unit tests were added
- [ ] **If feature added**: Added/extended example