Issue #1372 test regridder (#1463)
Fixes #1372

# Description
See linked issue for all my testing, I think xugrid's regridders are a
better replacement of ``imod.prepare.Regridder``! Time to report the
public! There is one edge case which I documented in the changelog.

- Extends tests to also verify the old Regridder works the same as
xugrids regridder. I didn't bother with a big refactor to add this as
separate tests, as I foresee that we will remove these unittests
altogether when the iMOD Python's Legacy Regridder class is removed.
- Extend documentation with additional pointers to xugrid's regridders.