#1480 Disable pulling the docker image team a build is started (#1481)
Fixes #1480

# Description
We are experiencing flaky builds on the TeamCity build server.

A theory is that the builds are failing due to pull collisions when the
docker images are pulled at the start of a build.
By disabling the 'explicit pull' option this might happen less frequent

With this change an image will only be pulled when its not present on
the agent. If the image is overwritten on the registry the old image has
to be manually removed from the agents
# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [ ] Links to correct issue
- [ ] Update changelog, if changes affect users
- [ ] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example