Issue #1487 improve readme (#1511)
Fixes #1487 

# Description
This is an attempt to update the README, to better showcase what iMOD
Python is capable of. I also added the code blocks from the README to an
example, so that we can test if the README examples still work. There is
one piece of the code which still doesn't work, due to #1510. That needs
to be resolved before merging this. For now, you can check if you agree
with the texts and example snippets. Also pay attention to whether I'm
fair enough to Flopy.

Update
--------
#1510 doesn't actually block this, I just had to drop some unexpected
coordinates. So this is ready for review.