Issue #305 Throw error upon space in directory name. (#1517)
Closes #305 

# Description

This throws an error when a user tries to write a model in a directory
with a space in it.

Follow-up to https://github.com/Deltares/imod-python/pull/1514

See my comment here:
https://github.com/Deltares/imod-python/pull/1514#issuecomment-2829863678

> Thanks for hotfixing this. FYI: This is not an immediate problem
anymore: I just talked to @TeunvanWoerkom, he moved to a drive without
spaces in the path.
> 
> I'm not sure if we should merge this, as the `wq` test bench doesn't
have the coverage of the `mf6` module. Most notably, we don't run a
model with iMOD-WQ to test if it breaks. The current `wq` module is
tried and tested in many projects, so that is sufficient to do work
with. I don't want to risk breaking it with changes.
> 
> Besides, Teun manually quoted the paths in his runfile to fix this and
then iMOD-WQ apparently also isn't able to deal with spaces in paths (at
least the version he used). It crashes upon attempting to create the
results directories.
> 
> Alternatively: We could raise an error if a user attempts to write a
`wq` model in a folder with a space in the path?


It turned out that adding quotes doesn't do the trick, as iMOD-WQ also
cannot handle spaces in paths, even when they are quoted in the runfile.
It crashes upon writing results.

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [x] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [x] Unit tests were added
- [ ] **If feature added**: Added/extended example