Issue #1519 weights in label arrays (#1520)
Fixes #1519 and #668

# Description
Implements the following:

- Provide weights to xugrid -> PyMetis calls
- Also call PyMetis for structured grids, instead of the naive uniform
splitter without options for weighting grids.
- Change name of label array to "label" instead of "idomain"
- Move ``get_label_array`` to ``imod.prepare.create_partition_labels``,
and add this to public API. Calling the old ``get_label_array`` results
in a DeprecationWarning, after which args are passed on to
``imod.prepare.create_partition_labels``.
- Fix bug where arrays provided to exchange creator do not have to be
named "idomain" anymore. They are now renamed to enforce a name.
- Pin xarray to 2025.3.0 to avoid error and HFB package with
\_\_repr\_\_

I was still on the fence about the latter, on one hand, I like that
these changes as they keep the code paths very similar and support
providing weights, on the other hand, the way Metis is now called, it
results in sloppy looking label arrays for structured grids.

For this [trivial
example](https://deltares.github.io/imod-python/examples/mf6/transport_2d.html)
with idomain equal to one everywhere. We had this:


![image](https://github.com/user-attachments/assets/b78f0748-8ade-4577-b691-1432a630ca89)

and with METIS we get this:


![image](https://github.com/user-attachments/assets/cba472a7-4293-485d-83b8-f9b89f0a2eb6)

That is still somewhat acceptable, though confusing as it looks sloppy,
so it deserves some extra explanation.
Perhaps manually create a simple label array here is better?

For the hondsrug example, which is not a perfect square model, METIS
accounts for the fact that the model domain is rectangular, which the
previous implementation does not:


![image](https://github.com/user-attachments/assets/ed77cef6-fdc6-4fb6-887f-f0808dff9b89)

So in the end I think it is an improvement