Issue #1564 reduce cost validation (#1565)
Fixes #1564

# Description
This extends the ``ValidationContext`` to also include a setting to trim
times during validation, just as the option added in #1562 to split.

I didn't update the changelog, as ``ValidationContext`` isn't part of
the public API yet, but I think it'd be useful to do that. Right now
users have to set a private attribute:

```
simulation._validation_context.ignore_time = True
``` 

I'll open up a follow-up issue for that. 

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [ ] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [x] Unit tests were added
- [ ] **If feature added**: Added/extended example