Issue #343 convert prj to mf6 docs (#1579)
Fixes #343

# Description
Adds example to import an iMOD5 model and regrid it to an unstructured
grid. There was quite some messing around going on to reallocate the
rivers. I think iMOD Python would benefit from a reallocate method, as
this example shows that it would save a lot of headaches, especially
after regridding. There is also an annoying thing with regridding where
a Ugrid2D object cannot be provided to the typedispatcher, where it
would be quite convenient for this example.

I'll open two issues for this.

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [ ] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [x] **If feature added**: Added/extended example