Issue #1373 add test plan (#1545)
Fixes #1373

# Description
Adds test plan for the 1.0 release. Two remarks:

- The LHM stress test information is still aspecific on where to
download data etc, as we are still conducting some tests here. We
haven't fully arrived at a final test case here (we are still tweaking
the case). When we have finished this work, we can upload it on a MinIO
bucket.
- The pixi task to generate a grid to be imported for qgis still has to
be made

UPDATE:
- I generated the pixi task to do the export for QGIS
- The user acceptance data (LHM stress test) job is now finished. I
versioned the data using DVC, so that changes to the data can be
committed to the git repo, without having to push the actual data to
Github. Instead, it's pushed/pulled from a Deltares MinIO storage.

# Checklist

- [x] Links to correct issue
- [ ] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [x] Unit tests were added
- [ ] **If feature added**: Added/extended example