Issue #996 Add Weekly CI job that runs the acceptance tests (#1673)
Fixes #988

# Description
This commit adds a weekly job that runs the acceptance tests.
A main "Test" job is created that is triggered every sunday. This job
triggers its dependencies.
For now that is only the "acceptance tests" job. However any future
weekly jobs can be attached to it.
I would also make sense to eventually put the "Pixi update" under this
job as that is a weekly job as well

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [ ] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example
- [ ] **If feature added**: Added feature to API documentation
- [ ] **If pixi.lock was changed**: Ran `pixi run generate-sbom` and
committed changes